pyproject.toml
[project]
name = "ysf-chan"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx[http2]>=0.27.2",
    "langchain-community>=0.3.2",
    "langchain-ollama>=0.2.0",
    "langchain>=0.3.3",
    "langgraph>=0.2.39",
    "python-dotenv>=1.0.1",
    "selenium>=4.25.0",
    "sqlite-vec>=0.1.3",
    "unstructured>=0.16.2",
    "vibrato>=0.2.0",
    "zstandard>=0.23.0",
]

[tool.uv]
dev-dependencies = [
    "pyright>=1.1.385",
    "ruff>=0.7.0",
]
